reorganized file structure, added integration test
