+++ crates/exchange/Cargo.toml
@@ -3,19 +3,24 @@ name = "exchange"
 version = "0.1.0"
 edition = "2021"
 
+[lib]
+name = "exchange"
+path = "src/lib.rs"
+
 [[bin]]
 name = "exchange-server"
-path = "src/server.rs"
+path = "src/bin/server.rs"
 
 [[bin]]
 name = "exchange-client"
-path = "src/client.rs"
+path = "src/bin/client.rs"
 
 [dependencies]
 tokio = { version = "1.0", features = ["rt-multi-thread", "macros"] }
 prost = "0.12"
 tonic = "0.10.2"
 alator = "0.3.3"
+tower = { version = "0.4" }
 
 [build-dependencies]
 tonic-build = "0.10"
